+++ simpleboto/__init__.py
@@ -1,6 +1,10 @@
+from simpleboto.athena import AthenaClient
+from simpleboto.logs import CLogger
 from simpleboto.s3 import S3Url, S3Client
 
 __all__ = [
+    'AthenaClient',
+    'CLogger',
     'S3Client',
     'S3Url'
 ]