+++ simpleboto/athena/constants/constants.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+"""
+(c) Charlie Collier, all rights reserved
+"""
+
+
+class C:
+    """
+    Constants class for Athena related constants.
+    """
+    S3_BUCKET = 'S3_BUCKET'
+    S3_PREFIX = 'S3_PREFIX'
+    FILE_FORMAT = 'FILE_FORMAT'
+    FILE_COMPRESSION = 'FILE_COMPRESSION'
+    PARTITION_SCHEMA = 'PARTITION_SCHEMA'
+    PARTITION_PROJECTION = 'PARTITION_PROJECTION'