+++ simpleboto/athena/utils/__init__.py
@@ -1,0 +1,29 @@
+from simpleboto.athena.utils.data_types import (
+    StringDType,
+    IntegerDType,
+    BigIntDType,
+    DoubleDType,
+    FloatDType,
+    BooleanDType,
+    DecimalDType,
+    VarCharDType,
+    TimestampDType,
+    DateDType,
+    DTypes
+)
+from simpleboto.athena.utils.schema import Schema
+
+__all__ = [
+    'StringDType',
+    'IntegerDType',
+    'BigIntDType',
+    'DoubleDType',
+    'FloatDType',
+    'BooleanDType',
+    'DecimalDType',
+    'VarCharDType',
+    'TimestampDType',
+    'DateDType',
+    'DTypes',
+    'Schema'
+]