+++ simpleboto/exceptions/custom_exceptions.py
@@ -10,7 +10,10 @@ class S3DelimiterError(Exception):
     """
     Exception class for S3 URLs with double slashes //.
     """
-    def __init__(self, url: str) -> None:
+    def __init__(
+        self,
+        url: str
+    ) -> None:
         super().__init__(f"The S3 URL {url} contains //")
 
 
@@ -28,3 +31,15 @@ class NoParameterError(Exception):
         super().__init__(
             f"Required parameter: {req_param} for {callable_}; current call: {callable_}({', '.join(arguments)})"
         )
+
+
+class InvalidSchemaType(Exception):
+    """
+    Exception class for specifying a data type not valid for a Schema.
+    """
+    def __init__(
+        self,
+        column: str,
+        dtype: str
+    ) -> None:
+        super().__init__(f"The data type {dtype} is not valid for Schema column {column}")