+++ tests/base_test.py
@@ -4,6 +4,7 @@
 """
 
 import os
+import shutil
 from unittest import TestCase
 
 import boto3
@@ -16,6 +17,15 @@ class BaseTest(TestCase):
         self.env_vars = {
             'REGION': 'us-east-1'
         }
+
+        self.test_data_dir = os.path.join(os.path.dirname(__file__), 'test_data')
+
+        self.tmp_dir = os.path.join(os.path.dirname(__file__), 'tmp')
+        if not os.path.exists(self.tmp_dir):
+            os.makedirs(self.tmp_dir)
+
+    def tearDown(self) -> None:
+        shutil.rmtree(self.tmp_dir)
 
     def _set_up_s3(self, bucket_name: str) -> None:
         self.s3c = boto3.client('s3', region_name=os.getenv('REGION'))